Standardizing on highp precision based on discussion in https://github.com/mapbox/mapbox-gl-js/issues/2096

+++ src/shaders/symbol_icon.vertex.glsl
@@ -6,9 +6,9 @@ attribute vec4 a_data;
 attribute vec3 a_size;
 uniform bool u_is_size_zoom_constant;
 uniform bool u_is_size_feature_constant;
-uniform mediump float u_size_t; // used to interpolate between zoom stops when size is a composite function
-uniform mediump float u_size; // used when size is both zoom and feature constant
-uniform mediump float u_layout_size; // used when size is feature constant
+uniform highp float u_size_t; // used to interpolate between zoom stops when size is a composite function
+uniform highp float u_size; // used when size is both zoom and feature constant
+uniform highp float u_layout_size; // used when size is feature constant
 uniform highp float u_camera_to_center_distance;
 uniform highp float u_pitch;
 uniform highp float u_collision_y_stretch;
@@ -19,7 +19,7 @@ uniform highp float u_collision_y_stretch;
 uniform mat4 u_matrix;
 
 uniform bool u_is_text;
-uniform mediump float u_zoom;
+uniform highp float u_zoom;
 uniform bool u_rotate_with_map;
 uniform vec2 u_extrude_scale;
 
@@ -35,11 +35,11 @@ void main() {
     vec2 a_offset = a_pos_offset.zw;
 
     vec2 a_tex = a_data.xy;
-    mediump vec2 label_data = unpack_float(a_data[2]);
-    mediump float a_labelminzoom = label_data[0];
-    mediump vec2 a_zoom = unpack_float(a_data[3]);
-    mediump float a_minzoom = a_zoom[0];
-    mediump float a_maxzoom = a_zoom[1];
+    highp vec2 label_data = unpack_float(a_data[2]);
+    highp float a_labelminzoom = label_data[0];
+    highp vec2 a_zoom = unpack_float(a_data[3]);
+    highp float a_minzoom = a_zoom[0];
+    highp float a_maxzoom = a_zoom[1];
 
     float size;
     // In order to accommodate placing labels around corners in
@@ -50,7 +50,7 @@ void main() {
     // currently rendered zoom level if text-size is zoom-dependent.
     // Thus, we compensate for this difference by calculating an adjustment
     // based on the scale of rendered text size relative to layout text size.
-    mediump float layoutSize;
+    highp float layoutSize;
     if (!u_is_size_zoom_constant && !u_is_size_feature_constant) {
         size = mix(a_size[0], a_size[1], u_size_t) / 10.0;
         layoutSize = a_size[2] / 10.0;
@@ -67,8 +67,8 @@ void main() {
 
     float fontScale = u_is_text ? size / 24.0 : size;
 
-    mediump float zoomAdjust = log2(size / layoutSize);
-    mediump float adjustedZoom = (u_zoom - zoomAdjust) * 10.0;
+    highp float zoomAdjust = log2(size / layoutSize);
+    highp float adjustedZoom = (u_zoom - zoomAdjust) * 10.0;
     // result: z = 0 if a_minzoom <= adjustedZoom < a_maxzoom, and 1 otherwise
     highp float z = 2.0 - step(a_minzoom, adjustedZoom) - (1.0 - step(a_maxzoom, adjustedZoom));
